+++ src/compat.ts
@@ -1,7 +1,6 @@
-import { promisify, callbackify, inherits } from 'util';
+import { callbackify, inherits } from 'util';
 import { EventEmitter } from 'events';
 import { Store as ExpressStore } from 'express-session';
-import { SessionStore } from './types';
 import MemoryStore from './store/memory';
 
 export function Store() {
@@ -28,12 +27,7 @@ expressSession.MemoryStore = CallbackMemoryStore;
 
 export { expressSession };
 
-export function promisifyStore(
-  store: Pick<ExpressStore, 'get' | 'destroy' | 'set'> & Partial<ExpressStore>
-): SessionStore {
-  store.get = promisify(store.get);
-  store.set = promisify(store.set);
-  store.destroy = promisify(store.destroy);
-  if (typeof store.touch === 'function') store.touch = promisify(store.touch);
-  return store as SessionStore;
+export function promisifyStore(store: ExpressStore): ExpressStore {
+  console.warn('promisifyStore has been deprecated! You can simply remove it.');
+  return store;
 }